backend/services/recording-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
SQLAlchemy>=2.0.20
psycopg[binary]>=3.2
python-dotenv>=1.0
google-auth>=2.18.0
google-api-python-client>=2.82.0

